refactor: subway에서 동일한 역에 대해 검증하던 로직을 서비스로 변경
